Optimize recomposition looping on Desktop and iOS (#14)
Currently calculateWindowSizeClass() will cause a recomposition every
time the window size changes. That is inefficient, and we really only
want to cause recompositions when the window size class changes.

This PR fixes this for Desktop and iOS. I've left the Android
implementation as upstream, so we don't diverge (even if it is
worse behavior).